[6.x] Split Console InteractsWithIO to external trait.
